Split Cookbook CSV into new challenge
This has been a constant source of confusion for students during the
Cookbook challenge. They just never really get the CSV part and it comes
up super early in the challenge so then they don't focus on the MVC
pattern, which is really the whole point of the challenge.

The solution I've come up with is to split the CSV part into a separate
challenge that they solve _only_ once they've already implemented the
MVC part. I've also provided much more guidance about how to do the CSV
bit and why the methods are structured the way they are.

This is still a work in progress since I wanted to get feedback on this
version before going a ahead and:
- [ ] Translating the `README` to non-English languages (I'll need some
  help on the French and Chinese but I can generate an initial draft
  with AI and my fake French that's just Spanish)
- [ ] Implementing the `fullstack-solutions` complement. This is mostly
  cutting and pasting from existing pieces, so it shouldn't be
  hard if we like the way the `README` is structured here.

Very curious to get your thoughts. Thankss so much!