Turn off eslint rules that are incompatible with importmaps
I removed import/extensions, which doesnt work with importmaps because
you have to say `.js` when importing with importmaps. resolve https://github.com/lewagon/fullstack-challenges/issues/2455

I also turned off import/no-unresolved, which is maybe more optional,
but I still think it's the right move. With importmaps, eslint doesn't
know which packages are imported in the HTML. So, unless you list them
as global variables when running the `eslint` command, then it thinks
any import of an external package is unresolved. In the Mapbox
challenge, we were adding mapboxgl as a global var for eslint. In the
Vue challenge, I noticed we weren't and so that challenge
would......always raise a style error I guess. Also if students _add_ a
package via importmaps at any point, they wouldn't know how to adjust
the eslint command. So I think it's better if we turn this style rule
off but just assume that students will see a console error anyway if
they imported the wrong file/package.