Update Mapbox geocoder API to v6
I guess we're technically still using the v5 JS library in the Geocoding
lecture when we implement it in Rails, but that caused us _soooo_ many
issues with importmap I don't want to change it. And it's quite
unrelated to the usecase here, so i don't think would confuse students.

resolve https://github.com/lewagon/fullstack-challenges/issues/2636